+++ resources/lambdaisland/cli/_licli.zsh
@@ -1,0 +1,17 @@
+#autoload
+
+_licli() {
+    local completion_output
+    completion_output="$(\
+        ${words[1]} __zsh_completions \
+        -- "${words[@]}" \
+    )"
+
+    if [ -n "$completion_output" ]; then
+        local -a pairs
+        pairs=("${(@f)${completion_output}}")
+        _describe "${words[1]}" pairs
+    fi
+
+    return 0
+}